factor out build options
